Move location models into separate file
